Reworked the SimCtlCommand – separated commands
